bump version and fix tome bug

+++ Assets/Scripts/Shop/ShopManager.cs
@@ -59,7 +59,7 @@ public class ShopManager : MonoBehaviour
             if (helper.Name == helperName && helper.DynamicCost <= Monitor.Influence)
             {
                 Monitor.Influence -= helper.DynamicCost;
-                helper.DynamicCost = (int) Math.Round(helper.DynamicCost * 1.3, 0);
+                helper.DynamicCost = (long) Math.Round(helper.DynamicCost * 1.3, 0);
                 // Monitor.Instance.UpdatePassiveIncomeText();
                 _audioManager.Play("CoinToss");
                 if (helper.AmountOwned == 0)
